Move AudioProcessor to common.
PiperOrigin-RevId: 483699606
