+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/BaseAudioProcessor.java
@@ -16,6 +16,7 @@
 package androidx.media3.exoplayer.audio;
 
 import androidx.annotation.CallSuper;
+import androidx.media3.common.audio.AudioProcessor;
 import androidx.media3.common.util.UnstableApi;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.nio.ByteBuffer;
