+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/SonicAudioProcessor.java
@@ -20,6 +20,7 @@ import static androidx.media3.common.util.Assertions.checkNotNull;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
+import androidx.media3.common.audio.AudioProcessor;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
