+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/audio/SilenceSkippingAudioProcessorTest.java
@@ -19,8 +19,8 @@ import static com.google.common.truth.Truth.assertThat;
 import static java.lang.Math.min;
 
 import androidx.media3.common.C;
+import androidx.media3.common.audio.AudioProcessor.AudioFormat;
 import androidx.media3.common.util.Assertions;
-import androidx.media3.exoplayer.audio.AudioProcessor.AudioFormat;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
