+++ libraries/transformer/src/main/java/androidx/media3/transformer/AudioTranscodingSamplePipeline.java
@@ -23,10 +23,10 @@ import static java.lang.Math.min;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
+import androidx.media3.common.audio.AudioProcessor;
+import androidx.media3.common.audio.AudioProcessor.AudioFormat;
 import androidx.media3.common.util.Util;
 import androidx.media3.decoder.DecoderInputBuffer;
-import androidx.media3.exoplayer.audio.AudioProcessor;
-import androidx.media3.exoplayer.audio.AudioProcessor.AudioFormat;
 import java.nio.ByteBuffer;
 import java.util.List;
 import org.checkerframework.checker.nullness.qual.RequiresNonNull;
