+++ libraries/transformer/src/main/java/androidx/media3/transformer/SpeedChangingAudioProcessor.java
@@ -19,8 +19,8 @@ package androidx.media3.transformer;
 import static java.lang.Math.min;
 
 import androidx.media3.common.C;
+import androidx.media3.common.audio.AudioProcessor;
 import androidx.media3.common.util.Util;
-import androidx.media3.exoplayer.audio.AudioProcessor;
 import androidx.media3.exoplayer.audio.BaseAudioProcessor;
 import androidx.media3.exoplayer.audio.SonicAudioProcessor;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
